Template renaming and performance improvement
