+++ v2/datastream-mongodb-to-firestore/src/main/resources/datastream-mongodb-to-mongodb-command-spec.json
